Flag to set repository from command line
New flag -r/--repository overrides any setting in the experiment config.
Allows the paths experiment.json to depend only on the layout of the
repository, making it easier to work with multiple checkouts of the same
code on the same or several computers, all at different paths.

The only caveat is that to pass this flag, experiment must be called as
'experiment run -r [source location]', since repository is only a flag
for the run command and not global.